AccEcn: encode and decode of ACE

+++ src/internet/model/tcp-accecn-data.h
@@ -32,11 +32,11 @@ public:
   void IniSenderCounters () { if(!m_isIniS) { m_isIniS = true; m_ecnCepS = 5; m_ecnE0bS = 1; m_ecnCebS = 0; m_ecnE1bS = 0;} }
   void IniReceiverCounters () { if(!m_isIniR) { m_isIniR = true; m_ecnCepR = 5; m_ecnE0bR = 1; m_ecnCebR = 0; m_ecnE1bR = 0;} }
 
-  uint8_t    m_ecnCepS    {0}; //!< For data sender, the number of packets marked respectively with the CE
+  uint32_t   m_ecnCepS    {0}; //!< For data sender, the number of packets marked respectively with the CE
   uint32_t   m_ecnCebS    {0}; //!< For data sender, the number of TCP payload bytes in packets marked respectively with the CE
   uint32_t   m_ecnE0bS    {0}; //!< For data sender, the number of TCP payload bytes in packets marked respectively with the ECT(0)
   uint32_t   m_ecnE1bS    {0}; //!< For data sender, the number of TCP payload bytes in packets marked respectively with the ECT(1)
-  uint8_t    m_ecnCepR    {0}; //!< For data receiver, the number of packets marked respectively with the CE
+  uint32_t   m_ecnCepR    {0}; //!< For data receiver, the number of packets marked respectively with the CE
   uint32_t   m_ecnCebR    {0}; //!< For data receiver, the number of TCP payload bytes in packets marked respectively with the CE
   uint32_t   m_ecnE0bR    {0}; //!< For data receiver, the number of TCP payload bytes in packets marked respectively with the ECT(0)
   uint32_t   m_ecnE1bR    {0}; //!< For data receiver, the number of TCP payload bytes in packets marked respectively with the ECT(1)
